+++ j2mod/src/main/java/com/smartgridready/driver/j2mod/J2ModModbusClientFactory.java
@@ -164,6 +164,35 @@ public class J2ModModbusClientFactory implements GenDriverAPI4ModbusFactory {
     }
 
     /**
+     * Creates a Modbus TCP transport.
+     * @param ipAddress the IP address or host name
+     * @param port the TCP port
+     * @param timeout the connection timeout in milliseconds
+     * @return a new instance of {@link J2ModModbusClient}
+     */
+    @Override
+    public GenDriverAPI4Modbus createTcpTransport(String ipAddress, int port, int timeout) {
+        return new J2ModModbusClient<>(
+            new ModbusTCPMaster(ipAddress, port, timeout, DEFAULT_RECONNECT, DEFAULT_RTU_OVER_TCP)
+        );
+    }
+
+    /**
+     * Creates a Modbus TCP transport.
+     * @param ipAddress the IP address or host name
+     * @param port the TCP port
+     * @param timeout the connection timeout in milliseconds
+     * @param rtuOverTcp uses RTU over a TCP connection, e.g. for IP-to-serial gateways
+     * @return a new instance of {@link J2ModModbusClient}
+     */
+    @Override
+    public GenDriverAPI4Modbus createTcpTransport(String ipAddress, int port, int timeout, boolean rtuOverTcp) {
+        return new J2ModModbusClient<>(
+            new ModbusTCPMaster(ipAddress, port, timeout, DEFAULT_RECONNECT, rtuOverTcp)
+        );
+    }
+
+    /**
      * Creates a Modbus UDP transport.
      * @param ipAddress the IP address or host name
      * @return a new instance of {@link J2ModModbusClient}